requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
django-elasticsearch-dsl
elasticsearch>=8.0.0,<9.0.0
pillow
openai>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0


